debug

+++ PoissonMixtureFit.py
@@ -81,7 +81,7 @@ def ObjLnPiQ(p, nm):
 		r += tf.reduce_sum(tf.gather(lnorderlist, kmatrix[j]))
 		r += scaler[j] * tf.cast(ietlist[j], tf.float64) * 10
 		tempr = tf.log(scaler[j]) + tf.log(tf.cast(ietlist[j], tf.float64))
-		r -= tempr * tf.reduce_sum(kmatrix[j])
+		r -= tempr * tf.cast(tf.reduce_sum(kmatrix[j]), tf.float64)
 	'''
 	before = list()
 	kmatrix = list()
